fix: 테이블 설계 수정 및 그에 따른 코드 리팩토링
* 적립금 테이블 따로 생성, 사용한 쿠폰 테이블 삭제, 쿠폰의 종류 Enum 매핑, 엔터티 연관관계 수정 등

test: 수정된 설계에 대해 테스트

docs: 설계 업데이트

+++ src/main/java/com/maetdori/buysomething/domain/Payment/PaymentRepository.java
@@ -2,5 +2,5 @@ package com.maetdori.buysomething.domain.Payment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Integer> {
 }
